requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
install==1.3.4
joblib==1.0.1
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
